+++ lzx/decoder/d16uncmp.c
@@ -8,7 +8,7 @@
 
 int NEAR decode_uncompressed_block(
     t_decoder_context * context,
-    long                bufpos,
+    int                bufpos,
     int                 amount_to_decode /* yes, it will equal 32768 */
 )
 {
@@ -56,10 +56,10 @@ bool NEAR handle_beginning_of_uncompressed_block(t_decoder_context *context)
     for (i = 0; i < NUM_REPEATED_OFFSETS; i++)
     {
         context->dec_last_matchpos_offset[i] =
-            ((ulong) *(  (byte *) context->dec_input_curpos)    )        |
-            ((ulong) *( ((byte *) context->dec_input_curpos) + 1) << 8)  |
-            ((ulong) *( ((byte *) context->dec_input_curpos) + 2) << 16) |
-            ((ulong) *( ((byte *) context->dec_input_curpos) + 3) << 24);
+            ((uint) *(  (byte *) context->dec_input_curpos)    )        |
+            ((uint) *( ((byte *) context->dec_input_curpos) + 1) << 8)  |
+            ((uint) *( ((byte *) context->dec_input_curpos) + 2) << 16) |
+            ((uint) *( ((byte *) context->dec_input_curpos) + 3) << 24);
 
         context->dec_input_curpos += 4; /* increment by 4 bytes */
     }
