+++ lzx/decoder/decout.c
@@ -7,7 +7,7 @@
 #include "decoder.h"
 
 
-void copy_data_to_output(t_decoder_context *context, long amount, const byte *data)
+void copy_data_to_output(t_decoder_context *context, int amount, const byte *data)
 {
     if (context->dec_output_buffer == NULL)
         return;
